Refactor: Extract VisualStateManager.cs co-located types into separate files
Move 5 non-primary types out of VisualStateManager.cs into their own files
in the same directory, following MAUI's one-type-per-file convention:

- VisualStateGroupList.cs  (was lines 167-401)
- VisualStateGroup.cs      (was lines 403-635)
- VisualState.cs           (was lines 637-741)
- VisualStateGroupListExtensions.cs  (was lines 743-796)
- WatchAddList.cs          (was lines 798-870)

VisualStateManager.cs is reduced from 871 to 166 lines and now contains
only the VisualStateManager static class. No logic changes; all types
remain in the Microsoft.Maui.Controls namespace.
