+++ src/Controls/src/Core/VisualStateManager.cs
@@ -163,709 +163,4 @@ namespace Microsoft.Maui.Controls
 			}
 		}
 	}
-
-	/// <summary>
-	/// A list of <see cref="VisualStateGroup"/> objects that enforces unique group and state names.
-	/// </summary>
-	public class VisualStateGroupList : IList<VisualStateGroup>
-	{
-		readonly IList<VisualStateGroup> _internalList;
-		internal bool IsDefault { get; private set; }
-
-		// Used to check for duplicate names; we keep it around because it's cheaper to create it once and clear it
-		// than to create one every time we need to validate
-		readonly HashSet<string> _names = new HashSet<string>(StringComparer.Ordinal);
-
-		void Validate(IList<VisualStateGroup> groups)
-		{
-			var groupCount = groups.Count;
-
-			// If we only have 1 group, no need to worry about duplicate group names
-			if (groupCount > 1)
-			{
-				_names.Clear();
-
-				// Using a for loop to avoid allocating an enumerator
-				for (int n = 0; n < groupCount; n++)
-				{
-					// HashSet will return false if the string is already in the set
-					if (!_names.Add(groups[n].Name))
-					{
-						throw new InvalidOperationException("VisualStateGroup Names must be unique");
-					}
-				}
-			}
-
-			// State names must be unique within this group list, so we'll iterate over all the groups
-			// and their states and add the state names to a HashSet; we throw an exception if a duplicate shows up
-
-			_names.Clear();
-
-			// Using nested for loops to avoid allocating enumerators
-			for (int groupIndex = 0; groupIndex < groupCount; groupIndex++)
-			{
-				// Cache the group lookup and states count; it's ugly, but it speeds things up a lot
-				var group = groups[groupIndex];
-				group.VisualElement = VisualElement;
-				group.UpdateStateTriggers();
-
-				var stateCount = group.States.Count;
-
-				for (int stateIndex = 0; stateIndex < stateCount; stateIndex++)
-				{
-					// HashSet will return false if the string is already in the set
-					if (!_names.Add(group.States[stateIndex].Name))
-					{
-						throw new InvalidOperationException("VisualState Names must be unique");
-					}
-				}
-			}
-		}
-
-		/// <summary>
-		/// Initializes a new instance of the <see cref="VisualStateGroupList"/> class.
-		/// </summary>
-		public VisualStateGroupList() : this(false)
-		{
-		}
-
-		/// <summary>
-		/// Initializes a new instance of the <see cref="VisualStateGroupList"/> class with the specified default state.
-		/// </summary>
-		/// <param name="isDefault">Indicates whether this list represents the default visual state groups.</param>
-		public VisualStateGroupList(bool isDefault)
-		{
-			IsDefault = isDefault;
-			_internalList = new WatchAddList<VisualStateGroup>(ValidateAndNotify);
-		}
-
-		void ValidateAndNotify(object sender, EventArgs eventArgs)
-		{
-			ValidateAndNotify(_internalList);
-		}
-
-		void ValidateAndNotify(IList<VisualStateGroup> groups)
-		{
-			if (groups.Count > 0)
-				IsDefault = false;
-
-			Validate(groups);
-			OnStatesChanged();
-		}
-
-		/// <inheritdoc />
-		public IEnumerator<VisualStateGroup> GetEnumerator()
-		{
-			return _internalList.GetEnumerator();
-		}
-
-		IEnumerator IEnumerable.GetEnumerator()
-		{
-			return ((IEnumerable)_internalList).GetEnumerator();
-		}
-
-		/// <inheritdoc />
-		public void Add(VisualStateGroup item)
-		{
-			if (item == null)
-			{
-				throw new ArgumentNullException(nameof(item));
-			}
-
-			_internalList.Add(item);
-
-			item.StatesChanged += ValidateAndNotify;
-		}
-
-		/// <inheritdoc />
-		public void Clear()
-		{
-			foreach (var group in _internalList)
-			{
-				group.StatesChanged -= ValidateAndNotify;
-			}
-
-			_internalList.Clear();
-		}
-
-		/// <inheritdoc />
-		public bool Contains(VisualStateGroup item)
-		{
-			return _internalList.Contains(item);
-		}
-
-		/// <inheritdoc />
-		public void CopyTo(VisualStateGroup[] array, int arrayIndex)
-		{
-			_internalList.CopyTo(array, arrayIndex);
-		}
-
-		/// <inheritdoc />
-		public bool Remove(VisualStateGroup item)
-		{
-			if (item == null)
-			{
-				throw new ArgumentNullException(nameof(item));
-			}
-
-			item.StatesChanged -= ValidateAndNotify;
-			return _internalList.Remove(item);
-		}
-
-		/// <inheritdoc />
-		public int Count => _internalList.Count;
-
-		/// <inheritdoc />
-		public bool IsReadOnly => false;
-
-		/// <inheritdoc />
-		public int IndexOf(VisualStateGroup item)
-		{
-			return _internalList.IndexOf(item);
-		}
-
-		/// <inheritdoc />
-		public void Insert(int index, VisualStateGroup item)
-		{
-			if (item == null)
-			{
-				throw new ArgumentNullException(nameof(item));
-			}
-
-			item.StatesChanged += ValidateAndNotify;
-			_internalList.Insert(index, item);
-		}
-
-		/// <inheritdoc />
-		public void RemoveAt(int index)
-		{
-			_internalList[index].StatesChanged -= ValidateAndNotify;
-			_internalList.RemoveAt(index);
-		}
-
-		public VisualStateGroup this[int index]
-		{
-			get => _internalList[index];
-			set => _internalList[index] = value;
-		}
-
-		WeakReference<VisualElement> _visualElement;
-		internal VisualElement VisualElement
-		{
-			get
-			{
-				if (_visualElement == null)
-					return null;
-				_visualElement.TryGetTarget(out var ve);
-				return ve;
-			}
-			set
-			{
-				_visualElement = new WeakReference<VisualElement>(value);
-			}
-		}
-
-		internal SetterSpecificity Specificity { get; set; }
-
-		void OnStatesChanged()
-		{
-			VisualElement?.ChangeVisualState();
-		}
-
-		public override bool Equals(object obj) => Equals(obj as VisualStateGroupList);
-		bool Equals(VisualStateGroupList other)
-		{
-			if (other is null)
-				return false;
-			if (Object.ReferenceEquals(this, other))
-				return true;
-			if (Count != other.Count)
-				return false;
-			for (var i = 0; i < Count; i++)
-				if (!this[i].Equals(other[i]))
-					return false;
-			return true;
-		}
-
-		public override int GetHashCode()
-		{
-			unchecked
-			{
-				var hash = 41;
-				for (var i = 0; i < Count; i++)
-					hash = (hash * 43) ^ this[i].GetHashCode();
-				return hash;
-			}
-		}
-
-	}
-
-	/// <summary>
-	/// Contains a collection of mutually exclusive <see cref="VisualState"/> objects and the setters to apply when transitioning between them.
-	/// </summary>
-	[RuntimeNameProperty(nameof(Name))]
-	[ContentProperty(nameof(States))]
-	public sealed class VisualStateGroup
-	{
-		/// <summary>
-		/// Initializes a new instance of the <see cref="VisualStateGroup"/> class.
-		/// </summary>
-		public VisualStateGroup()
-		{
-			States = new WatchAddList<VisualState>(OnStatesChanged);
-		}
-
-		/// <summary>
-		/// Gets or sets the type that this visual state group targets.
-		/// </summary>
-		public Type TargetType { get; set; }
-
-		/// <summary>
-		/// Gets or sets the name of the visual state group.
-		/// </summary>
-		public string Name { get; set; }
-
-		/// <summary>
-		/// Gets the collection of <see cref="VisualState"/> objects in this group.
-		/// </summary>
-		public IList<VisualState> States { get; }
-
-		/// <summary>
-		/// Gets the currently active <see cref="VisualState"/> in this group.
-		/// </summary>
-		public VisualState CurrentState { get; internal set; }
-
-		WeakReference<VisualElement> _visualElement;
-		internal VisualElement VisualElement
-		{
-			get
-			{
-				if (_visualElement == null)
-					return null;
-				_visualElement.TryGetTarget(out var ve);
-				return ve;
-			}
-			set
-			{
-				_visualElement = new WeakReference<VisualElement>(value);
-			}
-		}
-
-		internal VisualState GetState(string name)
-		{
-			foreach (VisualState state in States)
-			{
-				if (string.Equals(state.Name, name, StringComparison.Ordinal))
-				{
-					return state;
-				}
-			}
-
-			return null;
-		}
-
-		internal bool HasStateTriggers()
-		{
-			bool hasStateTriggers = false;
-
-			foreach (VisualState state in States)
-			{
-				if (state.StateTriggers.Count > 0)
-				{
-					hasStateTriggers = true;
-					break;
-				}
-			}
-
-			return hasStateTriggers;
-		}
-
-		internal VisualState GetActiveTrigger()
-		{
-			var defaultState = default(VisualState);
-			var visualState = defaultState;
-			var conflicts = new List<StateTriggerBase>();
-
-			for (var stateIndex = 0; stateIndex < States.Count; stateIndex++)
-			{
-				var state = States[stateIndex];
-				for (var triggerIndex = 0; triggerIndex < state.StateTriggers.Count; triggerIndex++)
-				{
-					var trigger = state.StateTriggers[triggerIndex];
-
-					if (trigger.IsActive)
-					{
-						if (visualState == defaultState)
-							visualState = state;
-
-						conflicts.Add(trigger);
-					}
-				}
-			}
-
-			if (conflicts.Count > 1)
-				visualState = ResolveStateTriggersConflict(conflicts);
-
-			return visualState;
-		}
-
-		VisualState ResolveStateTriggersConflict(List<StateTriggerBase> conflicts)
-		{
-			// When using StateTriggers to control visual states, the trigger engine uses the following rules to 
-			// score triggers and determine which trigger, and the corresponding VisualState, will be active:
-			//
-			// 1. Custom trigger that derives from StateTriggerBase
-			// 2. AdaptiveTrigger activated due to MinWindowWidth
-			// 3. AdaptiveTrigger activated due to MinWindowHeight
-			//
-			// If there are multiple active triggers at a time that have a conflict in scoring (i.e.two custom 
-			// triggers), then the first one declared in the markup file takes precedence.
-
-			var existCustomTriggers = conflicts.Where(c => !(c is AdaptiveTrigger));
-
-			if (existCustomTriggers.Count() > 1)
-			{
-				var firstExistCustomTrigger = existCustomTriggers.FirstOrDefault();
-				return firstExistCustomTrigger.VisualState;
-			}
-
-			var adaptiveTriggers = conflicts.Where(c => c is AdaptiveTrigger);
-
-			var minWindowWidthAdaptiveTriggers = adaptiveTriggers.Where(c => ((AdaptiveTrigger)c).MinWindowWidth != -1d).OrderByDescending(c => ((AdaptiveTrigger)c).MinWindowWidth);
-			var latestMinWindowWidthAdaptiveTrigger = minWindowWidthAdaptiveTriggers.FirstOrDefault();
-
-			if (latestMinWindowWidthAdaptiveTrigger != null)
-				return latestMinWindowWidthAdaptiveTrigger.VisualState;
-
-			var minWindowHeightAdaptiveTriggers = adaptiveTriggers.Where(c => ((AdaptiveTrigger)c).MinWindowHeight != -1d).OrderByDescending(c => ((AdaptiveTrigger)c).MinWindowHeight);
-			var latestMinWindowHeightAdaptiveTrigger = minWindowHeightAdaptiveTriggers.FirstOrDefault();
-
-			if (latestMinWindowHeightAdaptiveTrigger != null)
-				return latestMinWindowHeightAdaptiveTrigger.VisualState;
-
-			return default;
-		}
-
-		internal VisualStateGroup Clone()
-		{
-			var clone = new VisualStateGroup { TargetType = TargetType, Name = Name, CurrentState = CurrentState, VisualElement = VisualElement };
-
-			foreach (VisualState state in States)
-			{
-				state.VisualStateGroup = clone;
-				clone.States.Add(state.Clone());
-			}
-
-			if (VisualDiagnostics.IsEnabled && VisualDiagnostics.GetSourceInfo(this) is SourceInfo info)
-				VisualDiagnostics.RegisterSourceInfo(clone, info.SourceUri, info.LineNumber, info.LinePosition);
-
-			return clone;
-		}
-
-		internal void UpdateStateTriggers()
-		{
-			if (VisualElement == null)
-				return;
-
-			bool hasStateTriggers = HasStateTriggers();
-
-			if (!hasStateTriggers)
-				return;
-
-			var newStateTrigger = GetActiveTrigger();
-
-			if (newStateTrigger == null)
-				return;
-
-			var oldStateTrigger = CurrentState;
-
-			if (newStateTrigger == oldStateTrigger)
-				return;
-
-			VisualStateManager.GoToState(VisualElement, newStateTrigger.Name);
-		}
-
-		internal event EventHandler StatesChanged;
-
-		void OnStatesChanged(IList<VisualState> states)
-		{
-			if (states.Any(state => string.IsNullOrEmpty(state.Name)))
-			{
-				throw new InvalidOperationException("State names may not be null or empty");
-			}
-
-			foreach (var state in states)
-			{
-				state.VisualStateGroup = this;
-			}
-
-			StatesChanged?.Invoke(this, EventArgs.Empty);
-		}
-
-		public override bool Equals(object obj) => Equals(obj as VisualStateGroup);
-
-		bool Equals(VisualStateGroup other)
-		{
-			if (other is null)
-				return false;
-			if (object.ReferenceEquals(this, other))
-				return true;
-			if (Name != other.Name)
-				return false;
-			if (TargetType != other.TargetType)
-				return false;
-			if (States.Count != other.States.Count)
-				return false;
-			for (var i = 0; i < States.Count; i++)
-				if (!States[i].Equals(other.States[i]))
-					return false;
-			return true;
-		}
-
-		public override int GetHashCode()
-		{
-			unchecked
-			{
-				var hash = (Name, TargetType).GetHashCode();
-				for (var i = 0; i < States.Count; i++)
-					hash = (hash * 43) ^ States[i].GetHashCode();
-				return hash;
-			}
-		}
-	}
-
-	/// <summary>
-	/// Represents a named visual state with setters and triggers that define the appearance of a control.
-	/// </summary>
-	[RuntimeNameProperty(nameof(Name))]
-	public sealed class VisualState
-	{
-		/// <summary>
-		/// Initializes a new instance of the <see cref="VisualState"/> class.
-		/// </summary>
-		public VisualState()
-		{
-			Setters = new ObservableCollection<Setter>();
-			StateTriggers = new WatchAddList<StateTriggerBase>(OnStateTriggersChanged);
-		}
-
-		/// <summary>
-		/// Gets or sets the name of the visual state (e.g., "Normal", "Focused", "Disabled").
-		/// </summary>
-		public string Name { get; set; }
-
-		/// <summary>
-		/// Gets the collection of <see cref="Setter"/> objects that define property values for this state.
-		/// </summary>
-		public IList<Setter> Setters { get; }
-
-		/// <summary>
-		/// Gets the collection of <see cref="StateTriggerBase"/> objects that activate this state.
-		/// </summary>
-		public IList<StateTriggerBase> StateTriggers { get; }
-
-		/// <summary>
-		/// Gets or sets the type that this visual state targets.
-		/// </summary>
-		public Type TargetType { get; set; }
-		internal VisualStateGroup VisualStateGroup { get; set; }
-
-		internal VisualState Clone()
-		{
-			var clone = new VisualState { Name = Name, TargetType = TargetType };
-
-			foreach (var setter in Setters)
-			{
-				clone.Setters.Add(setter);
-			}
-
-			foreach (var stateTrigger in StateTriggers)
-			{
-				stateTrigger.VisualState = this;
-				clone.StateTriggers.Add(stateTrigger);
-			}
-
-			if (VisualDiagnostics.IsEnabled && VisualDiagnostics.GetSourceInfo(this) is SourceInfo info)
-				VisualDiagnostics.RegisterSourceInfo(clone, info.SourceUri, info.LineNumber, info.LinePosition);
-
-			return clone;
-		}
-
-		void OnStateTriggersChanged(IList<StateTriggerBase> stateTriggers)
-		{
-			foreach (var stateTrigger in stateTriggers)
-			{
-				stateTrigger.VisualState = this;
-			}
-
-			VisualStateGroup?.UpdateStateTriggers();
-		}
-
-		public override bool Equals(object obj) => Equals(obj as VisualState);
-
-		bool Equals(VisualState other)
-		{
-			if (other is null)
-				return false;
-			if (object.ReferenceEquals(this, other))
-				return true;
-			if (Name != other.Name)
-				return false;
-			if (TargetType != other.TargetType)
-				return false;
-			if (Setters.Count != other.Setters.Count)
-				return false;
-			if (StateTriggers.Count != other.StateTriggers.Count)
-				return false;
-			for (var i = 0; i < Setters.Count; i++)
-				if (!Setters[i].Equals(other.Setters[i]))
-					return false;
-			for (var i = 0; i < StateTriggers.Count; i++)
-				if (!StateTriggers[i].Equals(other.StateTriggers[i]))
-					return false;
-			return true;
-		}
-
-		public override int GetHashCode()
-		{
-			unchecked
-			{
-				var hash = (Name, TargetType).GetHashCode();
-				for (var i = 0; i < Setters.Count; i++)
-					hash = (hash * 43) ^ Setters[i].GetHashCode();
-				for (var i = 0; i < StateTriggers.Count; i++)
-					hash = (hash * 43) ^ StateTriggers[i].GetHashCode();
-				return hash;
-			}
-		}
-	}
-
-	internal static class VisualStateGroupListExtensions
-	{
-		internal static IList<VisualStateGroup> Clone(this IList<VisualStateGroup> groups)
-		{
-			var clone = new VisualStateGroupList();
-
-			foreach (var group in groups)
-			{
-				group.VisualElement = clone.VisualElement;
-				clone.Add(group.Clone());
-			}
-			
-			// Preserve specificity when cloning (issue #27202)
-			if (groups is VisualStateGroupList sourceList)
-			{
-				clone.Specificity = sourceList.Specificity;
-			}
-
-			if (VisualDiagnostics.IsEnabled && VisualDiagnostics.GetSourceInfo(groups) is SourceInfo info)
-				VisualDiagnostics.RegisterSourceInfo(clone, info.SourceUri, info.LineNumber, info.LinePosition);
-
-			return clone;
-		}
-
-		internal static bool HasVisualState(this VisualElement element, string name)
-		{
-			IList<VisualStateGroup> list = VisualStateManager.GetVisualStateGroups(element);
-			for (var i = 0; i < list.Count; i++)
-			{
-				VisualStateGroup group = list[i];
-				for (var j = 0; j < group.States.Count; j++)
-				{
-					if (group.States[j].Name == name)
-						return true;
-				}
-			}
-
-			return false;
-		}
-
-		internal static bool IsElementInSelectedState(this VisualElement element)
-		{
-			var groups = VisualStateManager.GetVisualStateGroups(element);
-			foreach (var group in groups)
-			{
-				if (group.CurrentState?.Name == VisualStateManager.CommonStates.Selected)
-				{
-					return true;
-				}
-			}
-
-			return false;
-		}
-	}
-
-	internal class WatchAddList<T> : IList<T>
-	{
-		readonly Action<List<T>> _onAdd;
-		readonly List<T> _internalList;
-
-		public WatchAddList(Action<List<T>> onAdd)
-		{
-			_onAdd = onAdd;
-			_internalList = new List<T>();
-		}
-
-		public IEnumerator<T> GetEnumerator()
-		{
-			return _internalList.GetEnumerator();
-		}
-
-		IEnumerator IEnumerable.GetEnumerator()
-		{
-			return ((IEnumerable)_internalList).GetEnumerator();
-		}
-
-		public void Add(T item)
-		{
-			_internalList.Add(item);
-			_onAdd(_internalList);
-		}
-
-		public void Clear()
-		{
-			_internalList.Clear();
-		}
-
-		public bool Contains(T item)
-		{
-			return _internalList.Contains(item);
-		}
-
-		public void CopyTo(T[] array, int arrayIndex)
-		{
-			_internalList.CopyTo(array, arrayIndex);
-		}
-
-		public bool Remove(T item)
-		{
-			return _internalList.Remove(item);
-		}
-
-		public int Count => _internalList.Count;
-
-		public bool IsReadOnly => false;
-
-		public int IndexOf(T item)
-		{
-			return _internalList.IndexOf(item);
-		}
-
-		public void Insert(int index, T item)
-		{
-			_internalList.Insert(index, item);
-			_onAdd(_internalList);
-		}
-
-		public void RemoveAt(int index)
-		{
-			_internalList.RemoveAt(index);
-		}
-
-		public T this[int index]
-		{
-			get => _internalList[index];
-			set => _internalList[index] = value;
-		}
-	}
 }
